pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "api-misuse-finder"
version = "0.1.0"
description = "Detect API misuse patterns in your codebase"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
keywords = ["api", "lint", "static-analysis", "code-quality"]
authors = [{ name = "takawasi" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.scripts]
api-finder = "api_finder.cli:main"

[project.urls]
Homepage = "https://github.com/takawasi/api-misuse-finder"
Repository = "https://github.com/takawasi/api-misuse-finder"

[tool.hatch.build.targets.wheel]
packages = ["api_finder"]

[tool.hatch.build.targets.sdist]
include = [
    "/api_finder",
    "/default_rules",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
